extract chart information in helper
